Unify annotation state helpers
Extract annotation-view and state-propagation helpers into a shared
internal module so constructs and primitives stay on the same
annotation-handling contract. This removes duplicated proxy and wrapper
logic that could drift on future fixes.

Add wrapper regressions for map(argument()) and group(command()), and
restore the changelog reference link so hongdown checks stay clean.

Refs https://github.com/dahlia/optique/issues/187
